Revert "Remove non-free! ga.js due to FSF and emacs-devel"
Turns out it is fine after all. Quote RMS:

"Our policy is that it is ok to link to a site that has some nonfree
Javascript code, as long as the page works even if the visitor blocks
execution of that code.  Not executing the Google Analytics code
won't make anything important fail, so ignore that please."